Added a guard for added negative amounts
Aggregates are consistency boundaries and ensures the consistency of the Domain. Right now it is possible to add -1000, which will store an MoneyAdded, which substracts money. This fix will make sure that negative amounts will always fire a MoneySubstracted and validate it's invariants. This can also be fixed in the user interface, but the whole point of an Aggregate is that is does not care about the outside world when it comes to it's own rules.